Big move around and re-write of stuff. Tested, seems working for raffles and ready to go. This is gonna be V1
