+++ .gitignore
@@ -1,1 +1,11 @@
+# PyCharm
 .idea
+api.key
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+# C extensions
+*.so
+# Jupyter Notebook
+.ipynb_checkpoints
