+++ requirements.txt
@@ -1,0 +1,4 @@
+# We want requests fo all the calls to discord, and back up to bitcoin and NIST
+requests
+# We use this to talk to discourse
+pydiscourse
