--- test/bikesheds.raffle
@@ -1,20 +1,0 @@
-RAFFLE_TIME="Tue 22 Dec 2020 12:56:00 PM PST"
-RAFFLE_NAME="Bike Sheds!"
-RAFFLE_COUNT="26"
-RAFFLE_DICE="THE DEMOCRATS DUMPED HUNDREDS OF THOUSANDS OF BALLOTS IN THE SWING STATES LATE IN THE EVENING. IT WAS A RIGGED ELECTION!!!"
-RAFFLE_SALT="0000000000000000000d16d93bab70e245c10eafc2bf376b47ab4b682136e9d7"
-ENTRANTS=$(cat <<EOF
-Bob
-Joe
-Guy
-Tom
-Tim
-Ann
-Jen
-Jeb
-Jax
-Kim
-Ray
-Han
-EOF
-);
