--- test/bikesheds.raffle.json
@@ -1,21 +1,0 @@
-{
-  "time": "Tue 22 Dec 2020 12:56:00 PM PST",
-  "name": "Bike Sheds!",
-  "count": 26,
-  "dice": "THE DEMOCRATS DUMPED HUNDREDS OF THOUSANDS OF BALLOTS IN THE SWING STATES LATE IN THE EVENING. IT WAS A RIGGED ELECTION!!!",
-  "salt": "0000000000000000000d16d93bab70e245c10eafc2bf376b47ab4b682136e9d7",
-  "entrants": [
-    "Bob",
-    "Joe",
-    "Guy",
-    "Tom",
-    "Tim",
-    "Ann",
-    "Jen",
-    "Jeb",
-    "Jax",
-    "Kim",
-    "Ray",
-    "Han"
-  ]
-}
