--- test/bikesheds.raffle.results
@@ -1,68 +1,0 @@
-RAFFLE_TIME="Tue 22 Dec 2020 12:56:00 PM PST"
-RAFFLE_NAME="Bike Sheds!"
-RAFFLE_COUNT="26"
-RAFFLE_DICE="THE DEMOCRATS DUMPED HUNDREDS OF THOUSANDS OF BALLOTS IN THE SWING STATES LATE IN THE EVENING. IT WAS A RIGGED ELECTION!!!"
-RAFFLE_SALT="0000000000000000000d16d93bab70e245c10eafc2bf376b47ab4b682136e9d7"
-ENTRANTS="Bob
-Joe
-Guy
-Tom
-Tim
-Ann
-Jen
-Jeb
-Jax
-Kim
-Ray
-Han"
-KEY="407d13b8d14cb2e00313c404116949a382c37fad786ed69aa9ee554edab17e92"
-DICE="8c5e0b8b89581a4b32e9f09a5d6176aa0962a302aa8293798e51f84c11c7b722"
-ROLL="542601ea1917126d183941ec92bd8feec44d185c00d11fac73e06b172f28e973"
-0="c92e78dd72e380d264fb9dc480830b8066dc9a7421acb2481e7a37c56be811ba"
-1="c267aedff7d23eedaf850a5ea0f709ebdccee53e17a2b7f45993e78f43816b52"
-2="e363624d96812fa31eb870797d0089e94293022051a7c7fd446b31adf4ff5339"
-3="cde2e14bc42723b8f1ddacb178d8ceb295038af93d54b7be85d8965ff31fcd6a"
-4="254c0bca74dd22db1267e4478c904f36c62ab87a3fac59b59bb88a45816e24cd"
-5="0d1b13239e65d7139c7fe0f5eac1f3c3692e60eb91b7d41f8258568afe00bfd5"
-6="133c47dec96a0becaa78e45efcac74924069d0b9b7d10f447729a3603839efe8"
-7="60b016fdab5502d4c170605248bf999386c71cf97f007a1b6cdc01bc6cf2357e"
-8="5680aa77f3e4b0b900e44492d9c09e5fd8971e1afbee851e390bec7485e0a792"
-9="348c710b7a079bfc2ce77a281391fdbf3672ee732eee29d694b744e42d14fcc1"
-10="261e4e843301d05fb1e147eb7ac39ad7e53f8155817da7fc217f5256eebfedaf"
-11="1571cf73a0f5f2c5b8aead7b825cc6f5f82e8f922a7fdeebed02342a00efc822"
-1571cf73a0f5f2c5b8aead7b825cc6f5f82e8f922a7fdeebed02342a00efc822="Han"
-0d1b13239e65d7139c7fe0f5eac1f3c3692e60eb91b7d41f8258568afe00bfd5="Joe"
-5680aa77f3e4b0b900e44492d9c09e5fd8971e1afbee851e390bec7485e0a792="Jeb"
-e363624d96812fa31eb870797d0089e94293022051a7c7fd446b31adf4ff5339="Tom"
-c92e78dd72e380d264fb9dc480830b8066dc9a7421acb2481e7a37c56be811ba="Jax"
-348c710b7a079bfc2ce77a281391fdbf3672ee732eee29d694b744e42d14fcc1="Kim"
-cde2e14bc42723b8f1ddacb178d8ceb295038af93d54b7be85d8965ff31fcd6a="Guy"
-133c47dec96a0becaa78e45efcac74924069d0b9b7d10f447729a3603839efe8="Tim"
-261e4e843301d05fb1e147eb7ac39ad7e53f8155817da7fc217f5256eebfedaf="Ray"
-60b016fdab5502d4c170605248bf999386c71cf97f007a1b6cdc01bc6cf2357e="Jen"
-254c0bca74dd22db1267e4478c904f36c62ab87a3fac59b59bb88a45816e24cd="Bob"
-c267aedff7d23eedaf850a5ea0f709ebdccee53e17a2b7f45993e78f43816b52="Ann"
-1571cf73a0f5f2c5b8aead7b825cc6f5f82e8f922a7fdeebed02342a00efc822="2"
-0d1b13239e65d7139c7fe0f5eac1f3c3692e60eb91b7d41f8258568afe00bfd5="2"
-5680aa77f3e4b0b900e44492d9c09e5fd8971e1afbee851e390bec7485e0a792="2"
-e363624d96812fa31eb870797d0089e94293022051a7c7fd446b31adf4ff5339="2"
-c92e78dd72e380d264fb9dc480830b8066dc9a7421acb2481e7a37c56be811ba="3"
-348c710b7a079bfc2ce77a281391fdbf3672ee732eee29d694b744e42d14fcc1="2"
-cde2e14bc42723b8f1ddacb178d8ceb295038af93d54b7be85d8965ff31fcd6a="2"
-133c47dec96a0becaa78e45efcac74924069d0b9b7d10f447729a3603839efe8="2"
-261e4e843301d05fb1e147eb7ac39ad7e53f8155817da7fc217f5256eebfedaf="2"
-60b016fdab5502d4c170605248bf999386c71cf97f007a1b6cdc01bc6cf2357e="2"
-254c0bca74dd22db1267e4478c904f36c62ab87a3fac59b59bb88a45816e24cd="2"
-c267aedff7d23eedaf850a5ea0f709ebdccee53e17a2b7f45993e78f43816b52="3"
-Tim="2"
-Joe="2"
-Han="2"
-Ann="3"
-Jax="3"
-Ray="2"
-Kim="2"
-Tom="2"
-Guy="2"
-Jen="2"
-Bob="2"
-Jeb="2"
